Simplify (*Test).removeNamespace
Use a = append(a[:i], a[i+1:]...) to delete the namespace at index i
from the Test.namespaces slice, also see
https://github.com/golang/go/wiki/SliceTricks

This avoids a slice out of bounds access if `DeleteNamespace`
(introduced in the next commit) is called for an arbitrary namespace.

Also add a test for addNamespace/removeNamespace to verify this works as
intended.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>