Fix check for running and ready pods
This commit fixes a scenario where WaitForPodsReady will block waiting
for the pods that are already running and ready, to be ready. However,
the number of ready and running pods are already greater than
`expectedReplicas`.

This is helpful when iterating on a test, rather than deleting the
entire cluster state for every test run.

Signed-off-by: Chris Tarazi <tarazichris@gmail.com>